Fix for failing tests on install
After upgrading DateTime-Format-Flexible from 0.33 to 0.34, a test
fails like this:

    t/03-parse.t ..... 1/?
    #   Failed test at t/03-parse.t line 32.
    #          got: 'America/New_York'
    #     expected: '-0500'
    # Looks like you failed 1 test of 18.
    t/03-parse.t ..... Dubious, test returned 1 (wstat 256, 0x100)

That's because new DateTime::Format::Flexible started to support
time zones enclosed in parenthesis and the teste string was
"Wed, 9 Nov 1994 09:50:32 -0500 (EST)".

I conclude that the test is broken because $DateTime->time_zone->name
is allowed to return a label instead of an offest. So I corrected the
test to compare an offset value.

The reason why this emerged now is that DateTimeX::Easy ignores almost
all DateTime::Format::Flexible return values because it plays with
time zones a bad way.